feat: agent optimization, terminator wiring, browser mode fixes & test coverage

+++ crates/sediman-tui/src/view/title_bar.rs
@@ -6,12 +6,10 @@ use super::messages::format_elapsed;
 pub fn render_title_bar(buf: &mut CellBuffer, area: Rect, app: &App) {
     let t = &app.theme;
 
-    // ── Fill background ──
     for sx in area.x..area.right() {
         buf.put_char(sx, area.y, ' ', Style::new().bg(t.background).fg(t.text));
     }
 
-    // ── Left: logo + spinner ──
     let spinner = if app.agent_running {
         format!(" {} ", app.spinner_char())
     } else {
@@ -29,7 +27,6 @@ pub fn render_title_bar(buf: &mut CellBuffer, area: Rect, app: &App) {
         buf.draw_str(vx, area.y, &version, Style::new().fg(t.text_muted).bg(t.background));
     }
 
-    // ── Right: provider/model + status ──
     let provider = &app.provider;
     let model = app.model.as_deref().unwrap_or("default");
 
@@ -40,7 +37,6 @@ pub fn render_title_bar(buf: &mut CellBuffer, area: Rect, app: &App) {
     };
     let status_color = if app.agent_running { t.success } else { t.text_muted };
 
-    // Provider pill
     let provider_pill = format!(" {} ", provider);
     let model_label = format!(" {} ", model);
     let sep = " \u{b7} ";
@@ -50,14 +46,13 @@ pub fn render_title_bar(buf: &mut CellBuffer, area: Rect, app: &App) {
     let mut rx = area.right().saturating_sub(right_w);
 
     buf.draw_str(rx, area.y, &provider_pill, Style::new().bg(t.secondary).fg(t.background));
-    rx += display_width(&provider_pill) as u16;
+    rx += display_width(&provider_pill);
     buf.draw_str(rx, area.y, &model_label, Style::new().bg(t.background_darker).fg(t.text));
-    rx += display_width(&model_label) as u16;
+    rx += display_width(&model_label);
     buf.draw_str(rx, area.y, sep, Style::new().fg(t.border_dim).bg(t.background));
     rx += display_width(sep);
     buf.draw_str(rx, area.y, &status, Style::new().fg(status_color).bg(t.background));
 
-    // ── Reconnecting warning ──
     if app.reconnecting {
         let warn = " \u{26a0} reconnecting ";
         let wx = area.x + display_width(&logo) + display_width(&version);
